Develop,emt: SkillItem Component integration
