+++ src/Components/SkillItem/SkillItem.module.css
@@ -1,8 +1,8 @@
-.skill-container {
+.skillContainer {
     display: inline-flex;
 }
 
-.tag-skill {
+.tagSkill {
     color: var(--gray-1, #393939);
     text-align: center;
     font-size: 16px;
